Fix crash: cast sensor window move() args to int
QWidget.move() rejects float arguments under PyQt5, so opening the sensor
window on script execution aborted with a TypeError. Cast the computed
coordinates to int, matching the existing int() usage elsewhere.

+++ MkECTL.py
@@ -774,8 +774,8 @@ class Ui(QMainWindow, IMainUI):
             self.sensorWindow.show()
         else:
             self.sensorWindow.show()
-            self.sensorWindow.move(geometry.width() / 2 - framesize.width() / 16,
-                                   geometry.height() / 2 - framesize.height() / 3)
+            self.sensorWindow.move(int(geometry.width() / 2 - framesize.width() / 16),
+                                   int(geometry.height() / 2 - framesize.height() / 3))
             self.sensorWindow_isOpen = True
         self.restoreDockWidget(self.sensorWindow)
         # self.subWindow.resize(QSize(909, 616))   # windowがfloatingしてるときはworkする。。
